+++ dev_requirements.txt
@@ -2,3 +2,4 @@ black==22.3.0
 mypy==0.961
 isort==5.10.1
 pytest==7.1.2
+autoflake==1.4
